fix: rename bitstream
